feat(graphql): rewrite faq, message, and promotion schemas/resolvers
Complete overhaul of FAQ, Message, and Promotion GraphQL modules.
Previous MessageResolver and PromotionResolver were stubs returning
empty lists and null. Now all three are fully functional with real
service calls.

FAQ (faq.graphqls + FAQResolver.java):
- Types: FAQ, FAQStats, HelpCategory, ContactOptions, FAQCategoryCount
- Public queries: faqs (with filter), faq, faqCategories, helpCategories,
  contactOptions
- Admin queries: adminFaqs, faqStats
- Admin mutations: createFAQ, updateFAQ, deleteFAQ, toggleFAQActive,
  reorderFAQs
- Updated input types: FAQCreateInput, FAQUpdateInput

Message (message.graphqls + MessageResolver.java):
- Types: Conversation, Message, ConversationStats
- Authenticated queries: myConversations, conversation, myMessageStats
- Admin queries: adminConversations, adminMessageStats
- Authenticated mutations: createConversation, replyToConversation,
  markConversationAsRead, toggleConversationStar, deleteConversation
- Admin mutations: adminReplyToConversation, updateConversationStatus,
  adminDeleteConversation
- FIXED: was returning empty lists/null, now calls MessageService

Promotion (promotion.graphqls + PromotionResolver.java):
- Types: AdminPromotion, SellerPromotion, AdminPromotionStats
- Admin queries: adminPromotions, activeAdminPromotions,
  adminPromotionStats
- Seller queries: sellerPromotions, activeSellerPromotions
- Admin mutations: createAdminPromotion, deleteAdminPromotion
- Seller mutations: createSellerPromotion, deleteSellerPromotion
- FIXED: was returning empty lists/null, now calls AdminPromotionService
  and SellerPromotionService

New/Updated input classes (6):
- FAQCreateInput, FAQUpdateInput, CreateConversationInput,
  SendMessageInput, AdminPromotionCreateInput, SellerPromotionCreateInput

Updated DTOs (3):
- AdminPromotionPage, SellerPromotionPage, ConversationConnection
  (fixed List<Object> to List<ConversationResponse>), FAQStats
  (Integer to Long)

+++ backend/src/main/java/ecommerce/graphql/dto/FAQStats.java
@@ -12,8 +12,8 @@ import lombok.Setter;
 @Setter
 @Builder
 public class FAQStats {
-    private Integer totalFAQs;
-    private Integer activeFAQs;
-    private Integer draftFAQs;
-    private Integer totalViews;
+    private Long totalFAQs;
+    private Long activeFAQs;
+    private Long draftFAQs;
+    private Long totalViews;
 }
